Allow empty passwords (#118)
* Allow empty passwords

* test

* remove user

* warning

* add empty password test

+++ integration/complex/passthrough_auth/run.sh
@@ -1,10 +1,11 @@
 #!/bin/bash
-set -ex
+set -e
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 export PGPASSWORD=pgdog
 export PGPORT=6432
 export PGHOST=127.0.0.1
 
+killall -TERM pgdog 2> /dev/null || true
 
 ${SCRIPT_DIR}/../../../target/release/pgdog \
     --config ${SCRIPT_DIR}/pgdog-enabled.toml \
@@ -17,6 +18,13 @@ if ! psql -U pgdog1 pgdog -c 'SELECT 1' > /dev/null; then
 fi
 
 psql -U pgdog pgdog -c 'SELECT 1' > /dev/null
+
+statement_timeout=$(psql -U pgdog1 pgdog -c 'SHOW statement_timeout' -t)
+
+if [[ "$statement_timeout" != *"100ms"* ]]; then
+    echo "AutoDB didn't pick up setting from users.toml"
+    exit 1
+fi
 
 killall -TERM pgdog
